requirement.txt
fastapi==0.115.0
uvicorn==0.30.6
motor==3.5.1
pymongo==4.8.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
itsdangerous==2.2.0
pydantic[email]==2.9.2
email-validator==2.2.0
python-multipart==0.0.9